Write into `txn` and `txn_participation` tables in parallel with other import procedures (#805)
* Write into txn and txn_participation tables in parallel with other import procedures.

* Change test.

* Add comments in idb.AddBlock().
